% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFCS.R
\name{concat_FCS}
\alias{concat_FCS}
\title{A function to concatenate flow cytometry data files (.FCS).}
\usage{
concat_FCS(fs, ev_max = 5000, ev_min = 0, verbose = TRUE)
}
\arguments{
\item{fs}{FlowSet with .FCS files with identical parameters to train the model.}

\item{ev_max}{Max number of events from each flowFrame.}

\item{ev_min}{Max number of events from each flowFrame.}

\item{verbose}{Verbosity.}
}
\value{
List with concatenated flowFrame, files, samples, ev_max, ev_min and verbose values if verbose = TRUE or only the concatenated flowFrame if verbose = FALSE.
}
\description{
A function to concatenate flow cytometry data files (.FCS) with the same parameters.
}
\keyword{FCS}
