% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFCS.R
\name{merge_FCS}
\alias{merge_FCS}
\title{A function to merge flow cytometry data files (.FCS).}
\usage{
merge_FCS(
  files = c(),
  common = c(1, 2, 3, 7, 10, 11),
  variable = c(4, 5, 6, 8, 9),
  normalize = TRUE,
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{files}{Vector of file names of the .FCS files to merge}

\item{common}{Vector of common columns number}

\item{variable}{Vector of variable columns number}

\item{normalize}{Quantile normalize the common parameters before imputation}

\item{impute}{logical. Missing values must be imputed?}

\item{verbose}{Verbosity}
}
\value{
Merged flowFrame
}
\description{
A function to merge flow cytometry data files (.FCS) and impute variable parameters using kNN on the common parameters.
}
\keyword{FCS}
